requirements.txt
# Core Django
Django==6.0.3
djangorestframework==3.15.2
django-cors-headers==4.9.0
django-redis

# Database & Env
psycopg2-binary==2.9.11
dj-database-url==3.1.2
python-dotenv==1.2.2


# Background Tasks (The New Stuff)
celery==5.4.0
redis==5.2.1

# Production & Static Files
gunicorn==25.1.0
whitenoise==6.12.0

# Helpers
asgiref==3.11.1
sqlparse==0.5.5
packaging==26.0


# Tests 
pytest-django==4.10.0
pytest-xdist==3.6.1      
model-bakery==1.20.3 